Improved WAIconButton with custom style; Renamed plugin file
