+++ src/WebAnnotationsIconButtonStyles.css
@@ -1,0 +1,38 @@
+#WebAnnoContainer {
+    margin-left: 7px;
+    margin-top: 8px;
+    cursor: pointer;   
+    color: gray;
+    font-family: 'Roboto', sans-serif;
+    padding: 11px 0px 8px 13px;
+    border-radius: 50%;
+    margin-right: 8px;
+}
+
+#WebAnnotContainter:hover {
+    background-color: lightslategray;
+}
+
+.tooltip .tooltiptext {
+    visibility: hidden;
+    width: 94px;
+    background-color: gray;
+    color: #fff;
+    text-align: center;
+    border-radius: 6px;
+    overflow: hidden;
+    white-space: nowrap;
+    padding: 5px 0;
+    font-size: 8pt;
+
+    /* Position the tooltip */
+    position: absolute;
+    z-index: 1;
+    top: 100%;
+    left: 50%;
+    margin-left: -26px;
+}
+  
+.tooltip:hover .tooltiptext {
+    visibility: visible;
+}
