+++ src/WebAnnotationsPlugin.js
